Created Java, DAOs, DTOs, Exceptions, Objects
I (Haroldas) have created the folders for the DAOs, DTOs, Exceptions and Objects, as well as placed the java code into each of the folders, for example EmployeeDaoInterface, MySqlDao and MySqlEmployeeDao in the DAOs folder.

Employee in DTOs folder, DaoException in Exceptions folder and App in the Objects folder.

Also created the CreateEmployee.sql, with all the necessary variables and values.