Added code for updating employee menu
Added code that allows the user to enter the id of the employee whose data they want to update. Then they can enter what data they want to change and what it will be changed to. Once they hit enter, the database will be updated with the new information
Fixed incorrect field names