Converting All Employees to Json String
Created JsonConverter class to convert all database information to a Json String. Takes all the information from the conncected database and stores it, then goes through each individual line of code and stores it as a string to be converted to Json, the Json is then parsed by the Gson into a more readable format with each line being displayed on a new line.
LocalDate cannot be serialised so I created the Adaptor class to take in the LocalDate for each employee's birth date and convert it into a string which can be serialised and it replaces the LocalDate date with the Json string date.
App - Added menu option to convert all employees to a Json String.
- displayJsonList calls the JsonConverter class and gets the information from the connected database. The list is gone through line by line and stored in individual employeeJson strings which are then connected to the converter to be converted to a Json format. Once one string is converted it will move to the next line of information and convert it. These steps will be repeated until it reaches the end of the database and all the strings are displayed.